OrderController + logic in viewing the order summaries

+++ eCom.Models/Coupon.cs
@@ -13,13 +13,13 @@ namespace eCom.Models
         public string Code { get; set; }
         public string Description { get; set; }
         public string DiscountType { get; set; } // e.g., "Percentage", "FixedAmount"
-        public decimal DiscountValue { get; set; }
+        public float DiscountValue { get; set; }
         public DateTime ExpirationDate { get; set; }
         public DateTime CreationDate { get; set; } = DateTime.Now;
         public int UsageLimit { get; set; }
         public int UsageCount { get; set; }
-        public decimal MinPurchaseAmount { get; set; }
-        public decimal MaxDiscountAmount { get; set; }
+        public float MinPurchaseAmount { get; set; }
+        public float MaxDiscountAmount { get; set; }
         public bool IsActive { get; set; }
         public bool IsSingleUse { get; set; }
         public virtual ICollection<Product> ApplicableProducts { get; set; } = new HashSet<Product>();
