+++ eCom.Models/ViewModels/CouponVM.cs
@@ -13,11 +13,11 @@ namespace eCom.Models.ViewModels
         public string Code { get; set; }
         public string Description { get; set; }
         public string DiscountType { get; set; } // e.g., "Percentage", "FixedAmount"
-        public decimal DiscountValue { get; set; }
+        public float DiscountValue { get; set; }
         public DateTime ExpirationDate { get; set; }
         public int UsageLimit { get; set; }
-        public decimal MinPurchaseAmount { get; set; }
-        public decimal MaxDiscountAmount { get; set; }
+        public float MinPurchaseAmount { get; set; }
+        public float MaxDiscountAmount { get; set; }
         public bool IsActive { get; set; }
         public bool IsSingleUse { get; set; }
         public List<Category>? ApplicableCategories { get; set; } = new List<Category>();
